requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
tinytuya
httpx
